requirements.txt
# Moodle AI Assistant - Dependencies
# Requires: Python 3.11 or 3.12 (recommended)

# Core
requests>=2.31.0
python-dotenv>=1.0.0

# LLM Providers (GLM + OpenAI only)
openai>=1.50.0

# Vector Store & Embeddings (no C++ compiler needed!)
faiss-cpu>=1.9.0
sentence-transformers>=3.0.0

# Document Processing
PyPDF2>=3.0.0
pymupdf>=1.24.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
markdownify>=0.13.0

# OCR (scanned PDF fallback)
pytesseract>=0.3.10
Pillow>=10.0.0

# Text Processing
tiktoken>=0.7.0
langchain-text-splitters>=0.2.0

# Web UI (optional — install separately: pip install gradio)
# gradio>=4.40.0

# Utilities
rich>=13.7.0
schedule>=1.2.0

# Telegram Bot
python-telegram-bot[job-queue]>=21.0
